requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.1/constraints-3.11.txt"

apache-airflow-providers-snowflake==5.7.0

psycopg2-binary
SQLAlchemy
pandas
pandera
astronomer-cosmos[dbt-postgres]
airflow-dbt-python